[#38982] renamed bcf vp wrapper
- removed unused code
